backend/requirements.txt
# FastAPI y dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0

# Validación
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Variables de entorno
python-dotenv==1.0.0

# CORS
python-cors==1.0.0

# Utilidades
python-dateutil==2.8.2

# Testing (opcional)
pytest==7.4.3
httpx==0.25.2
